changed text "parallel" and "series" to Enum in ComponentRelation
