+++ lcapy/solutionStep.py
@@ -1,5 +1,6 @@
 from lcapy import Circuit
 from lcapy import DrawWithSchemdraw
+from lcapy.componentRelation import ComponentRelation
 
 
 class SolutionStep:
@@ -8,7 +9,7 @@ class SolutionStep:
         self.cpt1 = step[1]
         self.cpt2 = step[2]
         self.newCptName = step[3]
-        self.relation = step[4]
+        self.relation: ComponentRelation = step[4]
         self.isInitialStep = not (step[1] or step[2] or step[3] or step[4])
         self.solutionText = None
         self.lastStep = None
